+++ keypoints/include/pcl/keypoints/impl/uniform_sampling.hpp
@@ -62,12 +62,12 @@ pcl::UniformSampling<PointInT>::detectKeypoints (PointCloudOut &output)
   pcl::getMinMax3D<PointInT>(*input_, min_p, max_p);
 
   // Compute the minimum and maximum bounding box values
-  min_b_[0] = (int)(floor (min_p[0] * inverse_leaf_size_[0]));
-  max_b_[0] = (int)(floor (max_p[0] * inverse_leaf_size_[0]));
-  min_b_[1] = (int)(floor (min_p[1] * inverse_leaf_size_[1]));
-  max_b_[1] = (int)(floor (max_p[1] * inverse_leaf_size_[1]));
-  min_b_[2] = (int)(floor (min_p[2] * inverse_leaf_size_[2]));
-  max_b_[2] = (int)(floor (max_p[2] * inverse_leaf_size_[2]));
+  min_b_[0] = static_cast<int> (floor (min_p[0] * inverse_leaf_size_[0]));
+  max_b_[0] = static_cast<int> (floor (max_p[0] * inverse_leaf_size_[0]));
+  min_b_[1] = static_cast<int> (floor (min_p[1] * inverse_leaf_size_[1]));
+  max_b_[1] = static_cast<int> (floor (max_p[1] * inverse_leaf_size_[1]));
+  min_b_[2] = static_cast<int> (floor (min_p[2] * inverse_leaf_size_[2]));
+  max_b_[2] = static_cast<int> (floor (max_p[2] * inverse_leaf_size_[2]));
 
   // Compute the number of divisions needed along all axis
   div_b_ = max_b_ - min_b_ + Eigen::Vector4i::Ones ();
@@ -90,9 +90,9 @@ pcl::UniformSampling<PointInT>::detectKeypoints (PointCloudOut &output)
         continue;
 
     Eigen::Vector4i ijk = Eigen::Vector4i::Zero ();
-    ijk[0] = (int)(floor (input_->points[(*indices_)[cp]].x * inverse_leaf_size_[0]));
-    ijk[1] = (int)(floor (input_->points[(*indices_)[cp]].y * inverse_leaf_size_[1]));
-    ijk[2] = (int)(floor (input_->points[(*indices_)[cp]].z * inverse_leaf_size_[2]));
+    ijk[0] = static_cast<int> (floor (input_->points[(*indices_)[cp]].x * inverse_leaf_size_[0]));
+    ijk[1] = static_cast<int> (floor (input_->points[(*indices_)[cp]].y * inverse_leaf_size_[1]));
+    ijk[2] = static_cast<int> (floor (input_->points[(*indices_)[cp]].z * inverse_leaf_size_[2]));
 
     // Compute the leaf index
     int idx = (ijk - min_b_).dot (divb_mul_);
@@ -119,7 +119,7 @@ pcl::UniformSampling<PointInT>::detectKeypoints (PointCloudOut &output)
 
   for (typename boost::unordered_map<size_t, Leaf>::const_iterator it = leaves_.begin (); it != leaves_.end (); ++it)
     output.points[cp++] = it->second.idx;
-  output.width = output.points.size ();
+  output.width = static_cast<uint32_t> (output.points.size ());
 }
 
 #define PCL_INSTANTIATE_UniformSampling(T) template class PCL_EXPORTS pcl::UniformSampling<T>;
